% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_formatting.R
\name{as_expression}
\alias{as_expression}
\title{Convert polynomial objects from \code{Momocs} into expressions.}
\usage{
as_expression(poly, expr = TRUE)
}
\arguments{
\item{poly}{An \code{OpnCoe} object generated by \code{Momocs::npoly()} or
\code{Momocs::opoly()}.}

\item{expr}{Indicate whether you want a string or expression returned.
Default is expression.}
}
\value{
an expression.
}
\description{
Convert polynomial objects from \code{Momocs} into expressions
for computing derivatives in R, eg \code{stats::deriv3()}.
}
\examples{

library(Momocs)

x <- 1:10
y <- x^2

# a fictional landmark dataset
mdat <- matrix(c(x,y), nrow=10, ncol=2)

# fit a second-order polynomial to landmarks
my_poly <- Momocs::npoly(mdat, 2)

# convert Momocs object into an expression
poly_express <- as_expression(my_poly)

# compute 1st (gradient) and 2nd (Hessian) derivatives
stats::deriv3(poly_express, "x")

}
\seealso{
\code{\link{parameterize}} for t-parameterizing Momocs polynomials
and \code{\link{as_function}} for creating R functions.
}
