% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_formatting.R
\name{as_param}
\alias{as_param}
\title{Parameterize polynomial objects from Momocs}
\usage{
as_param(poly)
}
\arguments{
\item{poly}{An OpnCoe object generated by \code{Momocs::npoly()} or \code{Momocs::opoly()}.}
}
\value{
a polynomial (e.g. y=x^2) function of the form
\code{function(t) {t, t^2}}.
}
\description{
Convert polynomial objects from Momocs into t-parameterized
polynomials readable by \code{pracma::arclength()}.
}
\examples{

library(Momocs)

x <- 1:10
y <- x^2

#a fictional landmark dataset
mdat <- matrix(c(x,y), nrow=10, ncol=2)

my_poly <- Momocs::npoly(mdat, 2)

as_param(my_poly)

}
\seealso{
\code{\link{as_expression}} for converting Momocs polynomials to expressions
and \code{\link{as_func}} for creating R functions.
}
