% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvature_circlefit.R
\name{curvature_circlefit}
\alias{curvature_circlefit}
\title{Calculate total curvature from a circle fitted to the landmarks}
\usage{
curvature_circlefit(landmark_matrix, x_range)
}
\arguments{
\item{landmark_matrix}{is a \code{matrix} object with \code{[,1]}
containing the x landmark coordinates and \code{[,2]} containing
the y landmark coordinates.}

\item{x_range}{the lower and upper x-value bounds to
calculate curvature. Concatenate the lower and upper bounds using
\code{c()}, E.g. for lower = 1 and upper = 10, type \code{c(1,10)}.}
}
\value{
a `list` with one named element. `$Ktot` is the total curvature in radians.
}
\description{
Calculate total curvature from a circle fitted to the landmarks
}
