% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_formatting.R
\name{express}
\alias{express}
\title{Convert polynomial objects from Momocs into expressions}
\usage{
express(poly, expr = TRUE)
}
\arguments{
\item{poly}{An OpnCoe object generated by \code{Momocs::npoly()} or \code{Momocs::opoly()}.}
}
\value{
an expression.
}
\description{
Convert polynomial objects from Momocs into expressions
that can be evaluated by e.g. \code{stats::deriv3()}.
}
\examples{
x <- 1:10
y <- x^2

#a fictional landmark dataset
mdat <- matrix(c(x,y), nrow=10, ncol=2)

my_poly <- Momocs::npoly(mdat, 2)

express(my_poly)

}
\seealso{
\code{\link{param}} for t-parameterizing Momocs polynomials
and \code{\link{func}} for creating R functions.
}
