% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_circlefit.R
\name{plot_circlefit}
\alias{plot_circlefit}
\title{Plot circle segment that has been fitted a landmark configuration}
\usage{
plot_circlefit(
  landmark_matrix,
  x_range,
  npoints = 1000,
  col = "red",
  pch = 1,
  ...
)
}
\arguments{
\item{landmark_matrix}{is a \code{matrix} object with \code{[,1]}
containing the x landmark coordinates and \code{[,2]} containing
the y landmark coordinates.}

\item{x_range}{the lower and upper x-value bounds to
plot the fitted circle. Concatenate the lower and upper bounds using
\code{c()}, E.g. for lower = 1 and upper = 10, type \code{c(1,10)}.}

\item{npoints}{a `numeric` stating how many points to sample from the fitted circle. Default is 1000.}

\item{col}{a `character` string indicating the colour to plot the
points from the fitted circle. Default is 'red'.}

\item{pch}{a `numeric` indicating the plotting symbol to be used.
Default is `1`. See ?graphics::points for details.}

\item{...}{further arguments passed to `graphics::points()`.}
}
\description{
Plot circle segment that has been fitted a landmark configuration
}
