% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_curvature.R
\name{total_curvature}
\alias{total_curvature}
\title{Calculate total curvature from polynomial functions
fitted to landmarked specimens.}
\usage{
total_curvature(Momocs_poly, x_range, subdiv = 1000, param = FALSE)
}
\arguments{
\item{Momocs_poly}{a polynomial \code{list} object
generated by \code{Momocs:::npoly.default()} or
\code{Momocs:::opoly.default()}. This function cannot yet
handle \code{OpnCoe} objects created by e.g. \code{Momocs:::npoly.Opn()}.}

\item{x_range}{the lower and upper x-value bounds to
calculate curvature. These are named 'baselines'
in \code{Momocs} polynomial objects.}

\item{subdiv}{number of subdivisions made
within \code{x_range}. Curvature is calculated at every
subdivision and summed to compute total curvature.
Default is 1000. The same number of subdivisions will be
applied to all curves, regardless of arclength.}
}
\value{
a list. The first element contains point-wise curvature. Units are
degrees per unit length. These values are useful for inspecting local
curvature features. The second element is total curvature. Units are radians.
Since arclength (ie specimen size) does not affect
the number of times point-wise curvature is computed (see \code{subdiv}),
there is no need to divide total curvature by arclength to adjust for specimen size.
}
\description{
Calculates point-wise
curvature along bounded polynomials fitted to curved
specimens. Polynomials are fitted to landmarked
specimens using package \code{Momocs}.
}
\examples{

library(Momocs)


# extract x bounds from Momocs polynomial object
x_range <- c(my_poly$baseline1[[1]], my_poly$baseline2[[1]])

# compute point-wise curvature 100 times
total_curvature(my_poly, x_range, 100)


}
