% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvature_spline.R
\name{curvature_spline}
\alias{curvature_spline}
\title{Calculate total curvature from smoothing  splines.}
\usage{
curvature_spline(landmark_matrix, x_range)
}
\arguments{
\item{x_range}{the lower and upper x-value bounds to
calculate curvature. Concatenate the lower and upper bounds using
\code{c()}, E.g. for lower = 1 and upper = 10, type \code{c(1,10)}.}

\item{model}{is a \code{lm} object fitted to the predicted values
of the first derivative of a \code{smooth.spline} object.
See examples below.}
}
\value{
a numeric vector length 1 indicating the total curvature in radians.
}
\description{
Calculate total curvature from smoothing  splines.
}
\examples{

}
