% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvature_tangle.R
\name{curvature_tangle}
\alias{curvature_tangle}
\title{Calculate total curvature by taking the derivative of
the Zahn and Roskies tangent angle function}
\usage{
curvature_tangle(landmark_matrix)
}
\arguments{
\item{landmark_matrix}{is a \code{matrix} object with \code{[,1]}
containing the x landmark coordinates and \code{[,2]} containing
the y landmark coordinates.}
}
\value{
a numeric vector length 1 indicating the total curvature in radians.
}
\description{
Calculate total curvature by taking the derivative of
the Zahn and Roskies tangent angle function
}
\examples{

}
