% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvature_tangle.R
\name{curvature_tangle}
\alias{curvature_tangle}
\title{Calculate total curvature by taking the derivative of
the Zahn and Roskies tangent angle function}
\usage{
curvature_tangle(coo_landmarks, interpolation = 500)
}
\arguments{
\item{function}{a single variable function of x.
E.g. for the polynomial function y=x^2 simply type \code{y^2}.}

\item{x_range}{the lower and upper x-value bounds to
calculate curvature. Concatenate the lower and upper bounds using
\code{c()}, E.g. for lower = 1 and upper = 10, type \code{c(1,10)}.}

\item{subdiv}{number of subdivisions made
within \code{x_range}. Curvature is calculated at every
subdivision and summed to compute total curvature.
Default is 1000. The same number of subdivisions will be
applied to all curves, regardless of arclength.}
}
\value{
a numeric vector length 1 indicating the total curvature in radians.
}
\description{
Calculate total curvature by taking the derivative of
the Zahn and Roskies tangent angle function
}
\examples{

}
