% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_formatting.R
\name{func}
\alias{func}
\title{Convert polynomial objects from Momocs into a function}
\usage{
func(poly, expr = TRUE)
}
\arguments{
\item{poly}{An OpnCoe object generated by \code{Momocs::npoly()} or \code{Momocs::opoly()}.}
}
\value{
a polynomial function (e.g. y = x^2) of the form \code{function(x) {x^2}}.
}
\description{
Convert polynomial objects from Momocs into a function
 readable by ...
}
\examples{
x <- 1:10
y <- x^2

#a fictional landmark dataset
mdat <- matrix(c(x,y), nrow=10, ncol=2)

my_poly <- Momocs::npoly(mdat, 2)

func(my_poly)
}
\seealso{
\code{\link{express}} for converting Momocs polynomials to expressions
and \code{\link{param}} for parameterizing by t.
}
